+++ CHANGELOG.md
@@ -1,9 +1,10 @@
-# 2.7.11 (UNRELEASED)
+# 2.7.11 (2021.06.25)
 
 * Update PHP and JS dependencies
 * Add ROLE_PERMANANT to be used for excluding accounts in user cleanup
 * Update ansible roles to fix playbook run
 * Add console command `app:users:list` with option to list inactive users
+* Dispatch AliasCreatedEvent after validation (Fixes: #216)
 
 # 2.7.10 (2021.05.20)